python/requirements.txt
bitalino==1.0
numpy==1.13.3
pyserial==3.4
